+++ CHANGELOG.rst
@@ -4,8 +4,8 @@
   Add xitrum.util.FileMonitor (requires Java 7)
 * `#308 <https://github.com/ngocdaothanh/xitrum/issues/308>`_
   Reload .po files updated at runtime
-* `#310 <https://github.com/ngocdaothanh/xitrum/issues/310>`_
-  Rename Json.generate to Json.serialize, Json.parse to Json.deserialize
+* `#311 <https://github.com/ngocdaothanh/xitrum/issues/311>`_
+  xitrum.util: Merge Json and SecureUrlSafeBase64 to SeriDeseri
 * `#306 <https://github.com/ngocdaothanh/xitrum/issues/306>`_
   Update Netty from 4.0.16 to 4.0.17